Cleanup asset parsers
