--- src/OpenSage.Game/Data/StreamFS/AssetReaders/AudioSettingsReader.cs
@@ -1,19 +1,0 @@
-﻿using System.IO;
-using OpenSage.Audio;
-
-namespace OpenSage.Data.StreamFS.AssetReaders
-{
-    public sealed class AudioSettingsReader : AssetReader
-    {
-        public override AssetType AssetType => AssetType.AudioSettings;
-
-        public override object Parse(Asset asset, BinaryReader reader, AssetImportCollection imports, AssetParseContext context)
-        {
-            var result = AudioSettings.ParseAsset(reader, asset);
-
-            context.AssetStore.AudioSettings.Current = result;
-
-            return result;
-        }
-    }
-}
