+++ src/OpenSage.Game/Graphics/Model.cs
@@ -1,10 +1,31 @@
-﻿using OpenSage.Content.Loaders;
+﻿using System.IO;
+using OpenSage.Content.Loaders;
 using OpenSage.Data.StreamFS;
+using OpenSage.Data.W3x;
+using OpenSage.FileFormats;
 
 namespace OpenSage.Graphics
 {
     public sealed class Model : BaseAsset
     {
+        internal static Model ParseAsset(BinaryReader reader, Asset asset, AssetImportCollection imports)
+        {
+            var hierarchy = imports.GetImportedData<ModelBoneHierarchy>(reader);
+            var subObjects = reader.ReadArrayAtOffset(() => W3xSubObject.Parse(reader, imports));
+
+            var modelSubObjects = new ModelSubObject[subObjects.Length];
+            for (var i = 0; i < subObjects.Length; i++)
+            {
+                var subObject = subObjects[i];
+                modelSubObjects[i] = new ModelSubObject(
+                    subObject.Name,
+                    hierarchy.Bones[subObject.BoneIndex],
+                    subObject.RenderObject);
+            }
+
+            return new Model(asset, hierarchy, modelSubObjects);
+        }
+
         public readonly ModelBoneHierarchy BoneHierarchy;
         public readonly ModelSubObject[] SubObjects;
